+++ src/main.rs
@@ -1,38 +1,36 @@
-mod util;
 mod term;
 mod arg_parser;
-mod jq;
 mod collection;
 
+use collection::{
+    jq::validate_jsons,
+    util::{DEFAULT_JSON_NAME, create_collection_dir, get_json_path, get_collection_dir_path}
+};
 use term::Term;
-use util::{DEFAULT_JSON_NAME, get_collection_dir_path, get_json_path};
+// use util::{DEFAULT_JSON_NAME, get_collection_dir_path, get_json_path};
 
 use std::path::Path;
 use std::env;
 
 
 fn main() {
-    // const COLLECTION_DIR_NAME: &str = ".memoire";
-    // let collection_dir_path: String = util::get_path_from_home_dir(COLLECTION_DIR_NAME);
-    // const DEFAULT_FILE_NAME: &str = "default.json";
-    // let default_json_path: String = format!("{}/{}", &collection_dir_path, DEFAULT_FILE_NAME);
-
     // Create Collection and default.json if does not exist
+    // TODO: Move this logic to collection
     if !Path::new(&get_json_path(DEFAULT_JSON_NAME)).exists() {
-        util::create_collection_dir(&get_collection_dir_path());
-        util::write_to_json(
-            &get_json_path(DEFAULT_JSON_NAME),
-            None
-        );
+        create_collection_dir(&get_collection_dir_path());
+        // util::write_to_json(
+        //     &get_json_path(DEFAULT_JSON_NAME),
+        //     None
+        // );
     }
 
     // Validate all json format
-    jq::validate_jsons(&get_collection_dir_path());
+    validate_jsons(&get_collection_dir_path());
 
     let mut term = Term::new();
     let args: Vec<String> = env::args().collect();
 
-    term.process_input(args);
+    // term.process_input(args);
     term.display();
